pyproject.toml
[project]
name = "course"
version = "0.1.0"
description = "GitHub documentation assistant for luongnv89/claude-howto"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "minsearch>=0.0.5",
    "numpy>=2.4.4",
    "openai>=1.108.2",
    "pydantic-ai==1.0.9",
    "python-dotenv>=1.2.2",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.39.0",
    "tqdm>=4.67.3",
]

[tool.uv]
dev-dependencies = []
